pyproject.toml
[project]
name = "agno-estudos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "agno>=2.2.10",
    "groq>=0.33.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.12",
]
